Show error if file exists

+++ src/Commands/FilePileInstallPile.php
@@ -119,10 +119,14 @@ class FilePileInstallPile extends Command
 
     private function installFile($pileBaseInstallPath, $file){
         $fullFilePath = $this->fullFilePath($pileBaseInstallPath,$file);
-        $this->info('+ Creating: '.$fullFilePath);
-        $fileContent = base64_decode($file->content);
-        $fileWriter = new FileWriter();
-        $fileWriter->create(base_path($fullFilePath),$fileContent);
+        if(file_exists(base_path($fullFilePath))){
+            $this->error('-File exists: '.$fullFilePath);
+        }else{
+            $this->info('+ Creating: '.$fullFilePath);
+            $fileContent = base64_decode($file->content);
+            $fileWriter = new FileWriter();
+            $fileWriter->create(base_path($fullFilePath),$fileContent);
+        }
     }
 
     private function showPostInstallationMessage($pile){
